remove MobDecorator interface, add restart, changed mobs and effects specs
