+++ 06-roguelike/src/main/kotlin/ru/spbau/maxim/mobs/effects/Effect.kt
@@ -1,7 +1,7 @@
 package ru.spbau.maxim.mobs.effects
 
-import ru.spbau.maxim.mobs.mobCore.MobDecoratorMutableImpl
+import ru.spbau.maxim.mobs.mobCore.MobDecoratorImpl
 import ru.spbau.maxim.mobs.mobCore.MobDecoratorMutable
 
-typealias EffectAbstract = MobDecoratorMutableImpl
+typealias EffectAbstract = MobDecoratorImpl
 typealias Effect = MobDecoratorMutable
