+++ 06-roguelike/src/main/kotlin/ru/spbau/maxim/mobs/effects/Frozen.kt
@@ -8,6 +8,6 @@ import ru.spbau.maxim.model.Model
  * Freezes: substitutes every mob action to Rest Action and mobCore does nothing
  */
 class Frozen: EffectAbstract() {
-    val timeActive = 3
+    val timeActive = 5
     override fun turn(env: Model): Action = Rest
 }
