+++ 06-roguelike/src/main/kotlin/ru/spbau/maxim/mobs/effects/Knife.kt
@@ -5,5 +5,5 @@ package ru.spbau.maxim.mobs.effects
  */
 class Knife: EffectAbstract() {
     override val attack: Int
-        get() = super.attack + 10
+        get() = super.attack + 8
 }
