--- 06-roguelike/src/main/kotlin/ru/spbau/maxim/mobs/mobCore/MobDecorator.kt
@@ -1,11 +1,0 @@
-package ru.spbau.maxim.mobs.mobCore
-
-/**
- * Interface for abstract mob decorator
- */
-interface MobDecorator: Mob {
-    /**
-     * return wrapped by current decorator mobCore
-     */
-    fun getInnerMob(): Mob
-}
