+++ 06-roguelike/src/main/kotlin/ru/spbau/maxim/mobs/mobCore/MobDecoratorMutable.kt
@@ -3,7 +3,7 @@ package ru.spbau.maxim.mobs.mobCore
 /**
  * Interface for MobDecorator, where you can change inner mob
  */
-interface MobDecoratorMutable: MobDecorator {
+interface MobDecoratorMutable: Mob {
     /**
      * change innerMob to mob
      */
